重构(build): 重构构建系统以支持自动生成应用名称和国际化
重构 CMake 构建系统，将应用名称从硬编码改为通过 CMake 变量自动生成
优化 Qt6 SQL 模块的依赖管理，集中到基础模块
新增配置文件位置迁移功能，支持在程序目录和 AppData/Local 之间切换

+++ cpp/modules/todo/todo_sync_server.cpp
@@ -240,7 +240,7 @@ void TodoSyncServer::推送下一个批次() {
     int startIndex = m_currentBatchIndex * maxBatchSize;
     int endIndex = std::min(startIndex + maxBatchSize, static_cast<int>(m_allUnsyncedItems.size()));
 
-    if (startIndex >= m_allUnsyncedItems.size()) {
+    if (startIndex >= static_cast<int>(m_allUnsyncedItems.size())) {
         // 所有批次都已推送完成
         qDebug() << "所有批次推送完成";
         setIsSyncing(false);
